Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.119.1",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "ruff>=0.14.1",
    "sqlmodel>=0.0.27",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "basedpyright>=1.32.0",
]
